backend/requirements.txt
# NuxAI Backend Dependencies

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Voice recognition (offline)
vosk==0.3.45
pyaudio==0.2.14

# Speech-to-text (v0.2)
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2

# Text-to-speech (v0.3)
pyttsx3==2.90

# Audio processing
numpy==1.26.2
soundfile==0.12.1
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Skills (v0.4)
aiohttp==3.9.1

# LLM Integration (v0.5)
gpt4all==2.0.2
chromadb==0.4.22

# Cross-platform (v0.6)
keyboard==0.13.5
pyperclip==1.8.2

# System Tray (v1.0)
pystray==0.19.5
pillow==10.1.0

# Web UI (v1.0)
jinja2==3.1.2

